Fixes set 5.2

+++ pyVoIP/VoIP/VoIP.py
@@ -975,8 +975,6 @@ class VoIPPhone:
                 self._callback_MSG_Bye(request)
             elif request.method == "CANCEL":
                 self._callback_MSG_Cancel(request)
-            elif request.method == "CANCEL":
-                self._callback_MSG_Cancel(request)
         else:
             # Only treat responses for INVITE as call-control.
             cseq = request.headers.get("CSeq", {})
@@ -1510,30 +1508,6 @@ class VoIPPhone:
         call.state = CallState.ENDED
         call._finalize_ended_call()
 
-    def _callback_MSG_Cancel(self, request: SIP.SIPMessage) -> None:
-        call_id = request.headers["Call-ID"]
-        debug(
-            request.summary(),
-            f"Inbound CANCEL call_id={call_id}",
-        )
-
-        call = self.calls.get(call_id)
-        if call is None or call.state != CallState.RINGING:
-            return
-
-        for rtp in call.RTPClients:
-            try:
-                rtp.stop()
-            except Exception:
-                pass
-
-        response = self.sip.gen_response(
-            call.request, SIP.SIPStatus.REQUEST_TERMINATED
-        )
-        self.sip.send_response(call.request, response)
-        call.state = CallState.ENDED
-        call._finalize_ended_call()
-
     def _callback_RESP_OK(self, request: SIP.SIPMessage) -> None:
         debug("OK recieved")
         call_id = request.headers["Call-ID"]
@@ -1855,7 +1829,11 @@ class VoIPPhone:
         to_delete = []
         for thread in self.threads:
             if not thread.is_alive():
-                call_id = self.threadLookup[thread]
+                call_id = self.threadLookup.get(thread)
+                if call_id is None:
+                    to_delete.append(thread)
+                    continue
+
                 call = self.calls.get(call_id)
                 if call is None:
                     debug("Unable to delete from calls dictionary!")
